For some reason I had to change the NaN data creation to not show a curve with datashader
Maybe this comes from some usage of dask in the data. I do not know.

The solution is a bit akward, since `data.values = np.NaN` did not
work. But the result is better now since the data to plot is only
three points which is faster for datashader to render (I guess...)